+++ js/main.js
@@ -1,80 +1,2 @@
-const NAMES = [
-  'Андрей',
-  'Елена',
-  'Дмитрий',
-  'Бадулай',
-  'Афанасий',
-  'Эдуард',
-  'Митрофан',
-];
-const PHOTO_DESCRIPTIONS =[
-  'Описание1',
-  'Описание2',
-  'Описание3',
-  'Описание4',
-  'Описание5',
-  'Описание6',
-  'Описание7',
-  'Описание8',
-  'Описание9',
-];
-
-const TEXT_FOR_COMMENT = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-//Функция, возвращающая случайное целое число
-function getRandomPositiveInteger(a, b) {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-//Получение рандомного элемента из переданного массива
-const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
-
-//Функция для получения чисел без повторений
-function getNumberWithoutRepeats(min,max){
-  const arrayRandomValues = [];
-  return function(){
-    let randomNumber = getRandomPositiveInteger(min,max);
-    while(arrayRandomValues.includes(randomNumber)){
-      randomNumber = getRandomPositiveInteger(min,max);
-    }
-    arrayRandomValues.push(randomNumber);
-    return randomNumber;
-  };
-}
-
-const genDescriptionId  = getNumberWithoutRepeats(1,25);
-const genImageAddressId = getNumberWithoutRepeats(1,25);
-const genCommentId = getNumberWithoutRepeats(1,999);
-
-//Функция создания комментария
-const createComment = ()=> ({
-  id:genCommentId(),
-  avatar:`img/avatar-${getRandomPositiveInteger(1,6)}.svg`,
-  message:getRandomArrayElement(TEXT_FOR_COMMENT),
-  name:getRandomArrayElement(NAMES)
-});
-
-const getSetComments = () => Array.from({length:2},createComment);
-
-// Функция создания объекта c информацией о фото
-const createDescriptionPhoto = () => ({
-  id:genDescriptionId(),
-  url:`photos/${genImageAddressId()}.jpg`,
-  description:getRandomArrayElement(PHOTO_DESCRIPTIONS),
-  likes:getRandomPositiveInteger(15,200),
-  comments:getSetComments(),
-});
-
-//Получение массива объектов
-// eslint-disable-next-line no-unused-vars
-const arrayObject = Array.from({length:25}, createDescriptionPhoto);
+import {arrayDescriptionsForPhoto} from'./data.js';
+arrayDescriptionsForPhoto();
